initialize and refactor
