+++ src/main/java/de/markherrmann/javauno/controller/TurnController.java
@@ -1,0 +1,9 @@
+package de.markherrmann.javauno.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping(value= "/turn")
+public class TurnController {
+}
