+++ src/main/java/de/markherrmann/javauno/data/state/component/GameLifecycle.java
@@ -1,4 +1,4 @@
-package de.markherrmann.javauno.data.state.components;
+package de.markherrmann.javauno.data.state.component;
 
 public enum GameLifecycle {
     SET_PLAYERS, RUNNING
