+++ src/main/java/de/markherrmann/javauno/data/state/component/game/Turn.java
@@ -1,0 +1,4 @@
+package de.markherrmann.javauno.data.state.component.game;
+
+class Turn {
+}
