+++ src/main/java/de/markherrmann/javauno/service/GameStateService.java
@@ -1,8 +1,8 @@
-package de.markherrmann.javauno.services;
+package de.markherrmann.javauno.service;
 
-import de.markherrmann.javauno.data.state.components.Game;
-import de.markherrmann.javauno.data.state.components.Player;
-import de.markherrmann.javauno.data.state.responses.GameState;
+import de.markherrmann.javauno.data.state.component.game.Game;
+import de.markherrmann.javauno.data.state.component.Player;
+import de.markherrmann.javauno.controller.response.GameState;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
