+++ src/main/java/de/markherrmann/javauno/service/TurnService.java
@@ -1,0 +1,7 @@
+package de.markherrmann.javauno.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class TurnService {
+}
